feat: extract Clix.Notification
